move DiffView files to new folder
